Separate fixture to conftest
