+++ tests/test_mysql_context_manager.py
@@ -3,19 +3,6 @@ from mysql_context_manager import __version__
 from mysql_context_manager import MysqlConnector
 
 
-@pytest.fixture()
-def mock_db(mocker):
-    mocker.patch("databases.Database.connect", return_value=None)
-    mocker.patch("databases.Database.disconnect", return_value=None)
-    mocker.patch(
-        "databases.Database.fetch_all",
-        return_value=[
-            [("username", "Aang")],
-        ],
-    )
-    mocker.patch("databases.Database.execute", return_value=None)
-
-
 def test_version():
     assert __version__ == "0.1.2"
